# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: cloud-asset-auditor
  description: |
    Multi-cloud inventory API. Audits Cloudflare, OCI, and Kubernetes
    into one canonical asset schema, plus a derived topology graph.

    All endpoints under `/api/v1/` are versioned per the project's
    contract (init-plan.md §6 invariant 6). `/healthz` and `/metrics`
    sit outside the versioned namespace because they're
    infrastructure concerns (probes + scrapers), not application API.
  version: "0.1.0"
  contact:
    name: cloud-asset-auditor
    url: https://github.com/cloud-auditor/cloud-asset-auditor
  license:
    name: TBD
    url: https://github.com/cloud-auditor/cloud-asset-auditor#license
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/providers:
    get:
      tags: [providers]
      summary: List registered providers
      responses:
        "200":
          description: Sorted list of registered provider names plus the server's auth mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvidersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/kube/contexts:
    get:
      tags: [providers]
      summary: List Kubernetes contexts from the server's kubeconfig
      description: |
        Returns the contexts defined in the server's kubeconfig (KUBECONFIG /
        ~/.kube/config) so a client can offer a "scan these clusters" picker.
        These name clusters the operator already configured — not credentials —
        and the audit endpoints validate any `kube_contexts` value against this
        same set.

        Always 200: an absent kubeconfig (or running in-cluster) returns an
        empty `contexts` list. A kubeconfig parse failure is reported in the
        `error` field.
      responses:
        "200":
          description: Available kube contexts plus the kubeconfig's current-context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KubeContextsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/audit:
    get:
      tags: [audit]
      summary: Streamed audit (SSE)
      description: |
        Server-Sent Events stream of one event per Asset. The full event sequence:

            event: meta        data: {"started_at": "RFC3339"}                          (exactly one, first)
            event: init_error  data: {"message": "..."}                                 (zero or more, before assets)
            event: asset       data: <Asset>                                            (zero or more, in arrival order)
            event: error       data: {"message": "..."}                                 (zero or more, interleaved)
            event: done        data: {"count": N, "elapsed_ms": M, "errors": K}        (exactly one, last)

        Long audits keep the connection open for minutes; use the
        `done` event (not the connection close) as the completion
        signal, since browsers may auto-reconnect on EventSource
        errors.
      parameters:
        - $ref: "#/components/parameters/ProvidersParam"
        - $ref: "#/components/parameters/TimeoutParam"
        - $ref: "#/components/parameters/KubeContextsParam"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEventStream"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/audit/export:
    get:
      tags: [audit]
      summary: Synchronous audit download
      description: |
        Runs the same audit as `/api/v1/audit` but blocks until done,
        then returns the rendered output as an attachment. Same data
        the CLI's `auditor audit` emits.
      parameters:
        - $ref: "#/components/parameters/ProvidersParam"
        - $ref: "#/components/parameters/TimeoutParam"
        - $ref: "#/components/parameters/KubeContextsParam"
        - name: format
          in: query
          description: Output format
          schema:
            type: string
            enum: [json, ndjson, csv, html, xlsx]
            default: json
        - name: sheet_by
          in: query
          description: >
            For `format=xlsx` only: how to partition assets across worksheets.
            One dimension — `none`, `provider`, `type`, `region`, `account`, or
            `tag:KEY` (e.g. `tag:compartment_id`) — or several joined with `+`
            for a composite (e.g. `region+tag:compartment_id` for one sheet per
            region/compartment, labelled `region (compartment)`). Ignored for
            other formats.
          schema:
            type: string
            default: provider
        - name: summary
          in: query
          description: >
            For `format=xlsx` only: when truthy (`1`/`true`/`yes`/`on`), prepend
            a "Summary" worksheet with totals and per-sheet/per-type counts,
            each per-sheet row linked to its worksheet. Ignored for other formats.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Rendered audit output
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="assets.json"'
            X-Auditor-Init-Errors:
              description: Semicolon-separated factory-time errors (e.g. missing provider creds)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
            application/x-ndjson:
              schema:
                type: string
                description: One JSON-encoded Asset per line
            text/csv:
              schema:
                type: string
                description: Header row plus one row per Asset; Tags flattened to k1=v1;k2=v2
            text/html:
              schema:
                type: string
                description: Self-contained audit report (same as `auditor audit -o html`)
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
                description: Multi-sheet Excel workbook; one worksheet per sheet_by group
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/topology:
    get:
      tags: [topology]
      summary: Inferred request-path graph between assets
      description: |
        Forces `--include-raw` internally so K8s resolvers can parse
        Ingress / HTTPRoute / Service payloads. The rendered output
        omits `raw` to stay readable.

        Resolvers:
          - `dns` — DNS records → matched LB/Service by IP or CNAME hostname (heuristic)
          - `lb-backend` — OCI LB IPs → K8s Service external IPs (heuristic)
          - `gateway-route` — K8s Ingress/HTTPRoute → backing Service (exact)
          - `service-backend` — K8s Service → selected Pods by label (exact)
          - `network-containment` — OCI subnet/gateway/OKE → VCN, NLB → subnet (exact)
          - `waf` — CF Rulesets/Access apps/Page Rules → protected zone (exact)
      parameters:
        - $ref: "#/components/parameters/ProvidersParam"
        - $ref: "#/components/parameters/TimeoutParam"
        - $ref: "#/components/parameters/KubeContextsParam"
        - name: hostname
          in: query
          description: Filter to the connected component reachable from this DNS hostname. Repeatable.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: include-orphans
          in: query
          description: Keep asset nodes that have no incident edges
          schema:
            type: boolean
            default: false
        - name: format
          in: query
          description: Output format. `json` returns the envelope below; others return a downloadable rendering.
          schema:
            type: string
            enum: [json, dot, mermaid, d2, graphml, excalidraw, drawio, html]
            default: json
        - name: group_by
          in: query
          description: |
            Cluster nodes by a dimension in the `dot`, `mermaid`, `d2`, and
            `drawio` renderers (ignored by the others). Empty or `none` renders flat.
            Also selects the bucketing dimension for `detail=medium|high`.
          schema:
            type: string
            enum: [provider, account, region, none]
        - name: detail
          in: query
          description: |
            Diagram detail level. `low` (default) renders every collected
            asset. `medium` collapses each group to one node per resource
            type; `high` collapses each group to a single node — the
            high-level network diagram. Both bucket by `group_by`, defaulting
            to `provider`. Collapsed edges carry a `count` multiplier, and
            edges that fall inside one bucket are reported as the node's
            `internal_edges` tag instead of as self-loops.
          schema:
            type: string
            enum: [low, medium, high]
            default: low
      responses:
        "200":
          description: Topology graph or rendered file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopologyResponse"
            text/vnd.graphviz:
              schema:
                type: string
            text/plain:
              schema:
                type: string
                description: Used by the `mermaid` format
            text/html:
              schema:
                type: string
                description: Used by the `html` format — a self-contained interactive viewer page
            application/graphml+xml:
              schema:
                type: string
                description: Used by the `graphml` format — XML for yEd / Gephi / Cytoscape
            application/xml:
              schema:
                type: string
                description: Used by the `drawio` format — mxGraph XML for draw.io / diagrams.net
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [topology]
      summary: Build a topology graph from client-supplied assets
      description: |
        Same graph engine as the GET variant, but the assets come from the
        request body instead of a fresh audit — instant, zero provider API
        calls. The web UI posts the assets it already streamed; CLI users can
        pipe a saved `auditor audit -o json --include-raw` snapshot. Include
        `raw` payloads on Kubernetes assets if you want the exact
        Ingress/HTTPRoute → Service edges (the resolvers read them).
      parameters:
        - name: hostname
          in: query
          description: Filter to the connected component reachable from this DNS hostname. Repeatable.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: include-orphans
          in: query
          description: Keep asset nodes that have no incident edges
          schema:
            type: boolean
            default: false
        - name: format
          in: query
          description: Output format. `json` returns the envelope below; others return a downloadable rendering.
          schema:
            type: string
            enum: [json, dot, mermaid, d2, graphml, excalidraw, drawio, html]
            default: json
        - name: group_by
          in: query
          description: |
            Cluster nodes by a dimension in the `dot`, `mermaid`, `d2`, and
            `drawio` renderers (ignored by the others). Empty or `none` renders flat.
            Also selects the bucketing dimension for `detail=medium|high`.
          schema:
            type: string
            enum: [provider, account, region, none]
        - name: detail
          in: query
          description: |
            Diagram detail level. `low` (default) renders every collected
            asset. `medium` collapses each group to one node per resource
            type; `high` collapses each group to a single node — the
            high-level network diagram. Both bucket by `group_by`, defaulting
            to `provider`. Collapsed edges carry a `count` multiplier, and
            edges that fall inside one bucket are reported as the node's
            `internal_edges` tag instead of as self-loops.
          schema:
            type: string
            enum: [low, medium, high]
            default: low
      requestBody:
        required: true
        description: |
          Either a bare JSON array of assets (what `auditor audit -o json`
          emits) or an `{"assets": [...]}` envelope. Bodies above 128 MiB
          are rejected with 413.
        content:
          application/json:
            schema:
              oneOf:
                - type: array
                  items:
                    $ref: "#/components/schemas/Asset"
                - type: object
                  properties:
                    assets:
                      type: array
                      items:
                        $ref: "#/components/schemas/Asset"
      responses:
        "200":
          description: Topology graph or rendered file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopologyResponse"
            text/vnd.graphviz:
              schema:
                type: string
            text/plain:
              schema:
                type: string
                description: Used by the `mermaid` format
            text/html:
              schema:
                type: string
                description: Used by the `html` format — a self-contained interactive viewer page
            application/graphml+xml:
              schema:
                type: string
                description: Used by the `graphml` format — XML for yEd / Gephi / Cytoscape
            application/xml:
              schema:
                type: string
                description: Used by the `drawio` format — mxGraph XML for draw.io / diagrams.net
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Request body exceeded the 128 MiB cap
  /api/v1/reach:
    get:
      tags: [topology]
      summary: Reachability analysis (fresh audit)
      description: |
        Answers "what can reach what" over the inferred topology. Runs a fresh
        audit with raw payloads (so the Kubernetes resolvers see Ingress /
        Service / NetworkPolicy specs), builds the graph, then traverses it.

        Edges point the way a request travels, so `from` walks forwards and
        `to` walks backwards. At least one of `from`, `to`, or `exposed` is
        required — an unconstrained reachability query has no meaningful answer.

        Absence of a path is **not** proof of isolation: it can equally mean no
        collected relationship joins the two assets.
      parameters:
        - $ref: "#/components/parameters/ProvidersParam"
        - $ref: "#/components/parameters/TimeoutParam"
        - $ref: "#/components/parameters/KubeContextsParam"
        - $ref: "#/components/parameters/ReachFromParam"
        - $ref: "#/components/parameters/ReachToParam"
        - $ref: "#/components/parameters/ReachExposedParam"
        - $ref: "#/components/parameters/ReachMaxHopsParam"
        - $ref: "#/components/parameters/ReachMaxPathsParam"
        - $ref: "#/components/parameters/ReachKindsParam"
        - $ref: "#/components/parameters/ReachIncludeDenyParam"
        - $ref: "#/components/parameters/TopologyFormatParam"
      responses:
        "200":
          description: Reachability result, or a rendered diagram of it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReachResult"
            text/plain:
              schema:
                type: string
        "400":
          description: No selector given, or a selector matched no assets
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [topology]
      summary: Reachability analysis (from supplied assets)
      description: |
        Same analysis as the GET form, but over assets supplied in the request
        body instead of a fresh audit. This is what the web UI calls: the
        Assets view already holds the streamed inventory, so re-running every
        provider to answer a reachability question would waste API quota.

        Note the body only carries `raw` if the *server* was started with
        `--include-raw`; without it the Kubernetes resolvers contribute no
        edges and paths through a cluster will be missed. Use the GET form,
        which forces raw on, when completeness matters.
      parameters:
        - $ref: "#/components/parameters/ReachFromParam"
        - $ref: "#/components/parameters/ReachToParam"
        - $ref: "#/components/parameters/ReachExposedParam"
        - $ref: "#/components/parameters/ReachMaxHopsParam"
        - $ref: "#/components/parameters/ReachMaxPathsParam"
        - $ref: "#/components/parameters/ReachKindsParam"
        - $ref: "#/components/parameters/ReachIncludeDenyParam"
        - $ref: "#/components/parameters/TopologyFormatParam"
      requestBody:
        required: true
        description: |
          Either a bare JSON array of assets or an `{"assets": [...]}`
          envelope. Bodies above 128 MiB are rejected.
        content:
          application/json:
            schema:
              oneOf:
                - type: array
                  items:
                    $ref: "#/components/schemas/Asset"
                - type: object
                  properties:
                    assets:
                      type: array
                      items:
                        $ref: "#/components/schemas/Asset"
      responses:
        "200":
          description: Reachability result, or a rendered diagram of it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReachResult"
        "400":
          description: Undecodable body, no selector given, or a selector matched no assets
        "413":
          description: Request body above the 128 MiB cap
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  parameters:
    ProvidersParam:
      name: providers
      in: query
      description: Comma-separated provider names (e.g. `cloudflare,kubernetes`). Empty selects every registered provider; the literal `none` selects zero.
      schema:
        type: string
        example: cloudflare,kubernetes
    TimeoutParam:
      name: timeout
      in: query
      description: Audit timeout as a Go duration string (e.g. `30m`, `2h`). Defaults to `10m`.
      schema:
        type: string
        pattern: '^\d+(\.\d+)?(ns|us|µs|ms|s|m|h)$'
        default: 10m
    KubeContextsParam:
      name: kube_contexts
      in: query
      description: >
        Comma-separated kubeconfig contexts for the Kubernetes provider to
        scan, or the literal `all` for every context. Each name is validated
        against the server's own kubeconfig (see `GET /api/v1/kube/contexts`);
        unknown names are dropped with an `init_error` / `X-Auditor-Init-Errors`
        warning. Empty leaves the server's default (current-context) in effect.
      schema:
        type: string
        example: prod-us,prod-eu
...
    ReachFromParam:
      name: from
      in: query
      description: >
        Source selector — a case-insensitive glob matched against each asset's
        id AND name. Walks edges forwards: "what can this reach?". Combine with
        `to` to enumerate routes between the two.
      schema:
        type: string
        example: api.example.com
    ReachToParam:
      name: to
      in: query
      description: >
        Target selector (same glob semantics as `from`). Walks edges backwards:
        "what can reach this?".
      schema:
        type: string
        example: "*postgres*"
    ReachExposedParam:
      name: exposed
      in: query
      description: >
        Start from the estate's public entry points — public DNS records,
        internet-facing load balancers, and Kubernetes Services a controller
        has actually published. Answers "what can the internet reach?".
      schema:
        type: boolean
        default: false
    ReachMaxHopsParam:
      name: max_hops
      in: query
      description: Maximum path length. Defaults to 6.
      schema:
        type: integer
        minimum: 1
        default: 6
    ReachMaxPathsParam:
      name: max_paths
      in: query
      description: >
        Maximum number of paths to return. Defaults to 25. The response sets
        `truncated` when the cap was hit — a truncated security result read as
        "these are all the ways in" is the dangerous kind of wrong.
      schema:
        type: integer
        minimum: 1
        default: 25
...
  responses:
    Unauthorized:
      description: Server is running with `--auth=basic` or `--auth=token` and the request didn't carry valid credentials.
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    BadRequest:
      description: Invalid query parameter (typically an unknown format)
      content:
        text/plain:
          schema:
            type: string
  schemas:
    ProvidersResponse:
      type: object
      required: [providers, auth_mode]
      properties:
        providers:
          type: array
          items: { type: string }
          example: [cloudflare, kubernetes, oci]
        auth_mode:
          type: string
          enum: [none, basic, token]
    KubeContextsResponse:
      type: object
      required: [contexts]
      properties:
        contexts:
          type: array
          items: { type: string }
          description: Context names from the server's kubeconfig (empty when none/in-cluster)
          example: [prod-us, prod-eu, staging]
        current:
          type: string
          description: The kubeconfig's current-context
          example: prod-us
        error:
          type: string
          description: Present only when the kubeconfig could not be parsed
    Asset:
      type: object
      description: The canonical provider-agnostic representation of a cloud resource.
      required: [provider, account_id, type, id, name]
      properties:
        provider:
          type: string
          enum: [cloudflare, oci, kubernetes]
        account_id:
          type: string
          description: Tenancy OCID / Cloudflare account ID / Kubernetes cluster name
        region:
          type: string
        type:
          type: string
          description: "Format: `<provider>.<resource>` (e.g. `cloudflare.dns_record`, `oci.compute.instance`); Kubernetes uses `<group>/<version>.<kind>` (e.g. `apps/v1.Deployment`, `v1.Pod`)"
          example: cloudflare.dns_record
        id:
          type: string
          description: Provider-native unique identifier
        name:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        tags:
          type: object
          additionalProperties:
            type: string
        raw:
          description: Full provider SDK payload — present only when `--include-raw` was requested
          oneOf:
            - type: object
            - type: array
            - { nullable: true }
...
    TopologyResponse:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items: { $ref: "#/components/schemas/Asset" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/Edge" }
        init_errors:
          type: array
          items: { type: string }
        errors:
          type: array
          items: { type: string }
    SSEEventStream:
      type: string
      description: |
        text/event-stream framed per RFC 8895. Each event has an
        `event:` line naming the type (`meta`, `init_error`, `asset`,
        `error`, `done`) and a `data:` line carrying a JSON object —
        see the endpoint description above for the per-event schema.
...